perf(cli): lazy-load command modules so light commands skip the heavy dep graph (#334)
Extracts `run`, `retry`/`abort`, and `clean` into separate modules loaded via
`await import()` from `cli.ts`. The dispatcher itself now only loads what the
invoked command actually needs.

Measured impact on `agent-ci --help`:

- Cold start: 240 ms → 20 ms
- Peak RSS: 88 MB → 42 MB

`--help` / unknown commands no longer load dockerode, @grpc/grpc-js, protobufjs,
ssh2, the runner graph, or the workflow parser. Every command behaves the same
otherwise. `--help`/`-h` now exits 0 (previously 1 — it had been falling through
the else branch).

Refs #334.
